reworked structure and rewrote classes
