+++ config.php
@@ -1,5 +1,9 @@
 <?php 
 
 return [
-
+    'db' => [
+        'host'=>'local',
+        'user'=>'admin',
+        'password'=>'admin123'
+    ]
 ];
