--- core/ApplicationTrait.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace core;
-
-trait ApplicationTrait
-{
-    private static $instance = null;
-
-    private function __construct()
-    {
-        //
-    }
-
-    private function __clone()
-    {
-        //
-    }
-
-    private function __wakeup()
-    {
-        //
-    }
-
-    public static function getInstance()
-    {
-        return static::$instance ?? (static::$instance = new static());
-    }
-}
