+++ core/Config.php
@@ -2,10 +2,17 @@
 
 namespace core;
 
+
 class Config
 {
-    public static function getConfig()
+
+    public static function getConfig(string $path)
     {
         $config = require_once('../config.php');
+        $arrPath = explode('/', $path);
+        // echo $config[$arrPath[0]][$arrPath[1]];
+        
     }   
 }
+
+Config::getConfig('db/host');
